requirements.txt
streamlit
pytesseract
pillow
pdf2image
opencv-python-headless
numpy